AVRO-2714: [C#] Optimize reading on .NET Core 2.1+
We can use stackallocs and ArrayPool to save on allocations when running on .NET Core.
