+++ lang/csharp/src/apache/main/Schema/Property.cs
@@ -121,7 +121,9 @@ namespace Avro
             int hash = this.Count;
             int index = 1;
             foreach (KeyValuePair<string, string> pair in this)
+#pragma warning disable CA1307 // Specify StringComparison
                 hash += (pair.Key.GetHashCode() + pair.Value.GetHashCode()) * index++;
+#pragma warning restore CA1307 // Specify StringComparison
             return hash;
         }
     }
